Split out some rendering stuff from OpenXRD3D11Kneeboard
