+++ src/injectables/OpenXRD3D11Kneeboard.h
@@ -22,6 +22,7 @@
 #include "OpenXRKneeboard.h"
 
 #include <OpenKneeboard/D3D11.h>
+#include <OpenKneeboard/D3D11/Renderer.h>
 #include <OpenKneeboard/SHM/D3D11.h>
 
 #include <OpenKneeboard/config.h>
@@ -62,25 +63,10 @@ class OpenXRD3D11Kneeboard final : public OpenXRKneeboard {
 
   winrt::com_ptr<ID3D11Device> mDevice;
   winrt::com_ptr<ID3D11DeviceContext> mImmediateContext;
-  std::unique_ptr<D3D11::SpriteBatch> mSpriteBatch;
+  std::unique_ptr<D3D11::Renderer> mRenderer;
 
-  PixelRect mSwapchainDimensions;
-
-  struct SwapchainBufferResources {
-    SwapchainBufferResources() = delete;
-    SwapchainBufferResources(
-      ID3D11Device*,
-      ID3D11Texture2D*,
-      DXGI_FORMAT renderTargetViewFormat);
-
-    ID3D11Texture2D* mTexture {nullptr};
-    winrt::com_ptr<ID3D11RenderTargetView> mRenderTargetView;
-  };
-
-  struct SwapchainResources {
-    PixelSize mDimensions;
-    std::vector<SwapchainBufferResources> mBufferResources;
-  };
+  using SwapchainBufferResources = D3D11::SwapchainBufferResources;
+  using SwapchainResources = D3D11::SwapchainResources;
 
   std::unordered_map<XrSwapchain, SwapchainResources> mSwapchainResources;
 };
